requirements.txt
cachetools==3.1.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.1
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
requests==2.27.1
rsa==4.0
six==1.12.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
uritemplate==4.1.1
urllib3==1.26.9
